Feature: loading stat info from file
This commit introduces the ability to storage the calculated information on data distributions.
User should present the path with key "stat_file" in the "training" section of the input file, by default "stat.npz".
If the train task is resuming from a checkpoint, the program will init the model parameters from scratch, and load params from file later.
If a checkpoint is not given, and the stat file exists, it will load those params. Currently, only one dataset is supported.
And if the file does not exist, the prolonged calculation process is executed, and the results are saved to the path of stat_file.

TODO: modify other models using se_atten