not using buffered dataloader for inferencing to avoid thread leak
The current implementation for inferencing creates a dataloader for each system.
The background consumer lacks the mechanism of exiting with unfinished tasks.
Hence, it will keep the iterator of dataloader, along with its workers, causing thread leak.
One possible approach is refactor the Tester class to construct loaders to all systems at once.
However, this will make output rmse and other results for each system impossible.
This commit takes the approach of removing the use of buffered iterator.